backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
boto3==1.28.62
python-docx==0.8.11